comments/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import Comment
from .serializers import CommentSerializer
from django_filters.rest_framework import DjangoFilterBackend


class CommentAPI(viewsets.ModelViewSet):
    queryset = Comment.objects.all().order_by('-id')
    serializer_class = CommentSerializer
    filterset_fields = ['author', 'post']
    filter_backends = [DjangoFilterBackend]
