requirements.txt
argcomplete==3.1.1
asgiref==3.7.2
boto3==1.28.11
botocore==1.31.11
certifi==2023.7.22
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==3.4.8
defusedxml==0.7.1
Django==4.0
django-cors-headers==4.2.0
django-filter==23.2
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
durationpy==0.5
hjson==3.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
placebo==0.9.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
zappa==0.57.0
